자신의 글만 삭제가능

+++ backend/src/main/java/com/example/backend/controller/board/BoardController.java
@@ -39,17 +39,28 @@ public class BoardController {
     }
 
     @DeleteMapping("delete/{id}")
-    public ResponseEntity<Map<String, Object>> delete(@PathVariable int id) {
-        if (service.remove(id)) {
-            return ResponseEntity.ok()
-                    .body(Map.of("message", Map.of("type", "success"
-                            , "text", STR."\{id}번 게시글이 삭제되었습니다.")));
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Map<String, Object>> delete(
+            @PathVariable int id,
+            Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            if (service.remove(id)) {
+                return ResponseEntity.ok()
+                        .body(Map.of("message", Map.of("type", "success"
+                                , "text", STR."\{id}번 게시글이 삭제되었습니다.")));
+            } else {
+                return ResponseEntity.internalServerError()
+                        .body(Map.of("message", Map.of("type", "error"
+                                , "text", "게시글 삭제 중 문제가 발생하였습니다.")));
+            }
+
         } else {
-            return ResponseEntity.internalServerError()
+            return ResponseEntity.status(403)
                     .body(Map.of("message", Map.of("type", "error"
-                            , "text", "게시글 삭제 중 문제가 발생하였습니다.")));
+                            , "text", "삭제 권한이 없습니다.")));
         }
     }
+
 
     @GetMapping("view/{id}")
     public Board view(@PathVariable int id) {
